+++ CHANGES.rst
@@ -6,7 +6,8 @@ Changes
 
 - Content disposition with semicolon in filename #917
 
-- Added `request_info` to response object and `ClientResponseError`.
+- Added `request_info` to response object and `ClientResponseError`. #1733
+
 
 2.0.1 (2017-03-21)
 ------------------